+++ android/src/main/java/com/reactnativeandroidwidget/RNWidgetProvider.java
@@ -1,18 +1,23 @@
 package com.reactnativeandroidwidget;
 
-import android.app.PendingIntent;
 import android.appwidget.AppWidgetManager;
 import android.appwidget.AppWidgetProvider;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.widget.RemoteViews;
 
-import com.facebook.react.HeadlessJsTaskService;
+import androidx.work.Data;
+import androidx.work.ExistingWorkPolicy;
+import androidx.work.OneTimeWorkRequest;
+import androidx.work.WorkManager;
+
 import com.facebook.react.bridge.Arguments;
+
+import org.json.JSONObject;
+
+import java.util.concurrent.TimeUnit;
 
 public class RNWidgetProvider extends AppWidgetProvider {
     @Override
@@ -20,9 +25,10 @@ public class RNWidgetProvider extends AppWidgetProvider {
         super.onAppWidgetOptionsChanged(context, appWidgetManager, appWidgetId, newOptions);
 
         if (isSizeChanged(context, appWidgetId)) {
-            Intent backgroundTaskIntent = buildIntent(context, appWidgetId, "WIDGET_RESIZED");
+            storeWidgetSize(context, appWidgetId);
 
-            executeJs(context, backgroundTaskIntent, appWidgetId);
+            Data data = buildData(context, appWidgetId, "WIDGET_RESIZED");
+            startBackgroundTask(context, data);
         }
     }
 
@@ -31,10 +37,11 @@ public class RNWidgetProvider extends AppWidgetProvider {
         super.onUpdate(context, appWidgetManager, appWidgetIds);
 
         for (int widgetId : appWidgetIds) {
+            String widgetAction = isWidgetNew(context, widgetId) ? "WIDGET_ADDED" : "WIDGET_UPDATE";
             storeWidgetSize(context, widgetId);
 
-            Intent backgroundTaskIntent = buildIntent(context, widgetId, "WIDGET_ADDED");
-            executeJs(context, backgroundTaskIntent, widgetId);
+            Data data = buildData(context, widgetId, widgetAction);
+            startBackgroundTask(context, data);
         }
     }
 
@@ -47,20 +54,17 @@ public class RNWidgetProvider extends AppWidgetProvider {
         }
 
         int widgetId = incomingIntent.getIntExtra("widgetId", -1);
-        String action = incomingIntent.hasExtra("widgetAction") ? incomingIntent.getStringExtra("widgetAction") : "WIDGET_CLICK";
 
-        if (action.equals("WIDGET_CLICK") && incomingIntent.hasExtra("clickAction")) {
-            switch (incomingIntent.getStringExtra("clickAction")) {
-                case "OPEN_APP":
-                    openApp(context);
-                    return;
-                case "OPEN_URI":
-                    openUri(context, incomingIntent.getBundleExtra("clickActionData").getString("uri"));
-                    return;
-            }
+        switch (incomingIntent.getStringExtra("clickAction")) {
+            case "OPEN_APP":
+                openApp(context);
+                break;
+            case "OPEN_URI":
+                openUri(context, incomingIntent.getBundleExtra("clickActionData").getString("uri"));
+                break;
+            default:
+                handleWidgetClick(context, incomingIntent, widgetId);
         }
-
-        emitAction(context, incomingIntent, widgetId, action);
     }
 
     @Override
@@ -70,76 +74,65 @@ public class RNWidgetProvider extends AppWidgetProvider {
         for (int widgetId : appWidgetIds) {
             removeWidgetSize(context, widgetId);
             RNWidgetCollectionService.deleteWidgetImages(context, widgetId);
+
+            Data data = buildData(context, widgetId, "WIDGET_DELETED");
+            startBackgroundTask(context, data);
         }
     }
 
-    private void emitAction(Context context, Intent incomingIntent, int widgetId, String action) {
-        Intent backgroundTaskIntent = buildIntent(context, widgetId, action);
-
-        if (incomingIntent.hasExtra("clickAction")) {
-            backgroundTaskIntent.putExtra("clickAction", incomingIntent.getStringExtra("clickAction"));
-            backgroundTaskIntent.putExtra("clickActionData", incomingIntent.getBundleExtra("clickActionData"));
-        }
-
-        startBackgroundTask(context, backgroundTaskIntent);
+    private Data buildData(Context context, int widgetId, String widgetAction, Data additionalData) {
+        return new Data.Builder()
+            .putString("widgetName", getClass().getSimpleName())
+            .putInt("widgetId", widgetId)
+            .putInt("width", RNWidgetUtil.getWidgetWidth(context, widgetId))
+            .putInt("height", RNWidgetUtil.getWidgetHeight(context, widgetId))
+            .putString("widgetAction", widgetAction)
+            .putAll(additionalData)
+            .build();
     }
 
-    private Intent buildIntent(Context context, int widgetId, String action) {
-        Bundle screenInfoBundle = new Bundle();
-        screenInfoBundle.putBundle("screenInfo", Arguments.toBundle(RNWidgetUtil.getScreenInfo(context)));
-
-        Intent backgroundTaskIntent = new Intent(context, RNWidgetBackgroundTaskService.class);
-        backgroundTaskIntent.putExtra("widgetName", getClass().getSimpleName());
-        backgroundTaskIntent.putExtra("widgetId", widgetId);
-        backgroundTaskIntent.putExtra("width", RNWidgetUtil.getWidgetWidth(context, widgetId));
-        backgroundTaskIntent.putExtra("height", RNWidgetUtil.getWidgetHeight(context, widgetId));
-        backgroundTaskIntent.putExtra("widgetAction", action);
-        backgroundTaskIntent.putExtras(screenInfoBundle);
-        return backgroundTaskIntent;
+    private Data buildData(Context context, int widgetId, String widgetAction) {
+        return buildData(context, widgetId, widgetAction, Data.EMPTY);
     }
 
-    private void executeJs(Context context, Intent backgroundTaskIntent, int widgetId) {
-        // Android 12+
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            addClickToUpdateOverlay(context, backgroundTaskIntent, widgetId);
-        } else {
-            startBackgroundTask(context, backgroundTaskIntent);
-        }
+    private void startBackgroundTask(Context context, Data data) {
+        workManagerWorkaround(context);
+
+        OneTimeWorkRequest headlessJsTaskWorkRequest =
+            new OneTimeWorkRequest.Builder(RNWidgetBackgroundTaskWorker.class)
+                .setInputData(data)
+                .build();
+
+        WorkManager
+            .getInstance(context)
+            .enqueue(headlessJsTaskWorkRequest);
     }
 
-    private void addClickToUpdateOverlay(Context context, Intent backgroundTaskIntent, int widgetId) {
-        RemoteViews remoteWidgetView = new RemoteViews(context.getPackageName(), R.layout.rn_widget);
-        RemoteViews clickableView = new RemoteViews(context.getPackageName(), R.layout.rn_widget_placeholder);
+    // `APPWIDGET_UPDATE` (`onUpdate`) method is called when the WorkManager queue is empty.
+    // Since we enqueue only on WorkRequest, the queue will be empty after every execution of the
+    // widgetTaskHandler.
+    // This is a bug in android (https://issuetracker.google.com/issues/115575872).
+    //
+    // The suggested workaround is to schedule another WorkRequest really far out into the future.
+    // So, we are creating a OneTimeWorkRequest with an initial delay of 10 years, with a name
+    // `app.package.WORK_MANAGER_HACK`.
+    //
+    // Every time when we schedule another WorkRequest, we REPLACE the WorkRequest with a new one.
+    // That way there is always at least one outstanding request, and `onUpdate` is not called when
+    // the WorkManager finishes with its work.
+    private void workManagerWorkaround(Context context) {
+        OneTimeWorkRequest headlessJsTaskWorkRequest =
+            new OneTimeWorkRequest.Builder(RNWidgetBackgroundTaskWorker.class)
+                .setInputData(Data.EMPTY)
+                .setInitialDelay(10 * 365, TimeUnit.DAYS)
+                .build();
 
-        Intent intent = new Intent(context.getPackageName() + ".WIDGET_CLICK");
-        intent.setClass(context, getClass());
-        intent.putExtra("widgetId", widgetId);
-        intent.putExtra("widgetAction", backgroundTaskIntent.getStringExtra("widgetAction"));
-
-        PendingIntent pendingIntent = PendingIntent.getBroadcast(
-            context,
-            (int) System.currentTimeMillis(),
-            intent,
-            PendingIntent.FLAG_CANCEL_CURRENT
-                | PendingIntent.FLAG_MUTABLE
-        );
-        clickableView.setOnClickPendingIntent(R.id.rn_widget_placeholder_click_to_update, pendingIntent);
-
-        remoteWidgetView.removeAllViews(R.id.rn_widget_clickable_container);
-        remoteWidgetView.addView(R.id.rn_widget_clickable_container, clickableView);
-
-        AppWidgetManager.getInstance(context)
-            .updateAppWidget(widgetId, remoteWidgetView);
-    }
-
-    private void startBackgroundTask(Context context, Intent serviceIntent) {
-        HeadlessJsTaskService.acquireWakeLockNow(context);
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            context.startForegroundService(serviceIntent);
-        } else {
-            context.startService(serviceIntent);
-        }
+        WorkManager.getInstance(context)
+            .enqueueUniqueWork(
+                context.getPackageName() + ".WORK_MANAGER_HACK",
+                ExistingWorkPolicy.REPLACE,
+                headlessJsTaskWorkRequest
+            );
     }
 
     private boolean isSizeChanged(Context context, int widgetId) {
@@ -150,18 +143,7 @@ public class RNWidgetProvider extends AppWidgetProvider {
         int newWidth = RNWidgetUtil.getWidgetWidth(context, widgetId);
         int newHeight = RNWidgetUtil.getWidgetHeight(context, widgetId);
 
-        boolean sizeChanged = oldWidth != newWidth || oldHeight != newHeight;
-
-        if (sizeChanged) {
-            SharedPreferences.Editor editor = sharedPref.edit();
-
-            editor.putInt(widgetId + "-width", newWidth);
-            editor.putInt(widgetId + "-height", newHeight);
-
-            editor.apply();
-        }
-
-        return sizeChanged;
+        return oldWidth != newWidth || oldHeight != newHeight;
     }
 
     private void storeWidgetSize(Context context, int widgetId) {
@@ -191,6 +173,13 @@ public class RNWidgetProvider extends AppWidgetProvider {
         editor.apply();
     }
 
+    private boolean isWidgetNew(Context context, int widgetId) {
+        SharedPreferences sharedPref = context.getSharedPreferences(
+            context.getPackageName() + ".WIDGET_SIZES", Context.MODE_PRIVATE);
+        int oldWidth = sharedPref.getInt(widgetId + "-width", -1);
+        return oldWidth == -1;
+    }
+
     private void openApp(Context context) {
         try {
             Intent launchIntent = context.getPackageManager().getLaunchIntentForPackage(context.getPackageName());
@@ -208,5 +197,20 @@ public class RNWidgetProvider extends AppWidgetProvider {
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
+
+    private void handleWidgetClick(Context context, Intent incomingIntent, int widgetId) {
+        Data additionalData = Data.EMPTY;
+
+        if (incomingIntent.hasExtra("clickAction")) {
+            Bundle clickActionData = incomingIntent.getBundleExtra("clickActionData");
+            additionalData = new Data.Builder()
+                .putString("clickAction", incomingIntent.getStringExtra("clickAction"))
+                .putString("clickActionData", new JSONObject(Arguments.fromBundle(clickActionData).toHashMap()).toString())
+                .build();
+        }
+
+        Data data = buildData(context, widgetId, "WIDGET_CLICK", additionalData);
+        startBackgroundTask(context, data);
     }
 }