+++ docs/sidebars.js
@@ -66,10 +66,6 @@ const sidebars = {
       type: 'doc',
       id: 'demo',
     },
-    {
-      type: 'doc',
-      id: 'limitations',
-    },
   ],
 
   api: [