+++ example/src/App.tsx
@@ -9,6 +9,7 @@ import { ExampleScreens, ListScreen } from './screens/ListScreen';
 import { SvgScreen } from './screens/SvgScreen';
 import { TextScreen } from './screens/TextScreen';
 import { ClickDemoWidgetPreviewScreen } from './screens/widget-preview/ClickDemoWidgetPreviewScreen';
+import { DebugEventsWidgetPreviewScreen } from './screens/widget-preview/DebugEventsWidgetPreviewScreen';
 import { FitnessWidgetPreviewScreen } from './screens/widget-preview/FitnessWidgetPreviewScreen';
 import { ListDemoWidgetPreviewDeepLinkScreen } from './screens/widget-preview/ListDemoWidgetPreviewDeepLinkScreen';
 import { ListDemoWidgetPreviewScreen } from './screens/widget-preview/ListDemoWidgetPreviewScreen';
@@ -86,6 +87,11 @@ export function App() {
           name="ListDemoWidgetPreviewDeepLinkScreen"
           component={ListDemoWidgetPreviewDeepLinkScreen}
           options={{ title: 'List Widget Deep Link' }}
+        />
+        <Stack.Screen
+          name="DebugEventsWidgetPreviewScreen"
+          component={DebugEventsWidgetPreviewScreen}
+          options={{ title: 'Debug Widget Events' }}
         />
       </Stack.Navigator>
     </NavigationContainer>