+++ example/src/widgetTaskHandler.tsx
@@ -3,6 +3,10 @@ import React from 'react';
 import type { WidgetTaskHandlerProps } from 'react-native-android-widget';
 import { ClickDemoWidget } from './widgets/ClickDemoWidget';
 import { CounterWidget } from './widgets/CounterWidget';
+import {
+  DEBUG_EVENTS_STORAGE_KEY,
+  DebugEventsWidget,
+} from './widgets/DebugEventsWidget';
 import { FitnessWidget } from './widgets/FitnessWidget';
 import { ListDemoWidget } from './widgets/ListDemoWidget';
 import { ResizableMusicWidget } from './widgets/ResizableMusicWidget';
@@ -17,6 +21,7 @@ const nameToWidget = {
   Counter: CounterWidget,
   ClickDemo: ClickDemoWidget,
   List: ListDemoWidget,
+  DebugEvents: DebugEventsWidget,
 };
 
 const COUNTER_STORAGE_KEY = 'CounterWidget:count';
@@ -27,6 +32,19 @@ export async function widgetTaskHandler(props: WidgetTaskHandlerProps) {
   const Widget = nameToWidget[
     widgetInfo.widgetName as keyof typeof nameToWidget
   ] as any;
+
+  if (widgetInfo.widgetName === 'DebugEvents') {
+    let events = await writeAndGetEvents(
+      widgetInfo.widgetId,
+      props.widgetAction
+    );
+    if (props.widgetAction === 'WIDGET_CLICK') {
+      AsyncStorage.setItem(DEBUG_EVENTS_STORAGE_KEY, '[]');
+      events = [];
+    }
+    props.renderWidget(<DebugEventsWidget events={events} />);
+    return;
+  }
 
   switch (props.widgetAction) {
     case 'WIDGET_RESIZED':
@@ -45,6 +63,19 @@ export async function widgetTaskHandler(props: WidgetTaskHandlerProps) {
       } else {
         props.renderWidget(<Widget {...widgetInfo} />);
       }
+      break;
+
+    case 'WIDGET_UPDATE':
+      if (widgetInfo.widgetName === 'Counter') {
+        const count = +((await AsyncStorage.getItem(COUNTER_STORAGE_KEY)) ?? 0);
+        props.renderWidget(<CounterWidget count={count} />);
+      } else {
+        props.renderWidget(<Widget {...widgetInfo} />);
+      }
+      break;
+
+    case 'WIDGET_DELETED':
+      // Do nothing
       break;
 
     case 'WIDGET_CLICK':
@@ -77,4 +108,15 @@ export async function widgetTaskHandler(props: WidgetTaskHandlerProps) {
     default:
       break;
   }
+}
+
+async function writeAndGetEvents(
+  widgetId: number,
+  action: string
+): Promise<string[]> {
+  const data = await AsyncStorage.getItem(DEBUG_EVENTS_STORAGE_KEY);
+  const items = JSON.parse(data ?? '[]');
+  items.push(`Widget #${widgetId} ${action} - ${new Date().toLocaleString()}`);
+  AsyncStorage.setItem(DEBUG_EVENTS_STORAGE_KEY, JSON.stringify(items));
+  return items;
 }